@@ -15,17 +15,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.1",
-    "@angular/cdk": "~13.0.1",
-    "@angular/common": "~13.0.1",
-    "@angular/compiler": "~13.0.1",
-    "@angular/core": "~13.0.1",
-    "@angular/forms": "~13.0.1",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^13.0.1",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
     "@angular/material": "^13.0.1",
-    "@angular/platform-browser": "~13.0.1",
-    "@angular/platform-browser-dynamic": "~13.0.1",
-    "@angular/router": "~13.0.1",
-    "@angular/service-worker": "~13.0.1",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@apidevtools/json-schema-ref-parser": "^9.0.7",
     "@apollo/client": "^3.0.0",
     "@auth0/angular-jwt": "^5.0.2",
@@ -55,14 +55,14 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^13.0.0",
-    "@angular-devkit/build-angular": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.12",
     "@angular-eslint/builder": "12.6.1",
     "@angular-eslint/eslint-plugin": "12.6.1",
     "@angular-eslint/eslint-plugin-template": "12.6.1",
     "@angular-eslint/schematics": "12.6.1",
     "@angular-eslint/template-parser": "12.6.1",
-    "@angular/cli": "~13.0.2",
-    "@angular/compiler-cli": "~13.0.1",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^14.3.0",
     "@ngrx/schematics": "^12.5.1",
     "@types/jasmine": "~3.6.0",
     "@types/lodash-es": "^4.17.4",
@@ -83,7 +83,7 @@
     "ng2-charts": "^3.1.2",
     "prettier": "^2.8.7",
     "pretty-quick": "^3.1.3",
-    "typescript": "~4.4.4",
+    "typescript": "~4.8.4",
     "webpack-bundle-analyzer": "^4.4.1",
     "webpack-sources": "^3.2.1"
   }
